backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
databases
pydantic
python-dotenv
httpx